move hyperoptimizer out as a module; move gpu config to util
